feat(adapter-node): add http2 support
