+++ .vscode/settings.json
@@ -1,5 +1,9 @@
 {
 	"typescript.tsdk": "node_modules/typescript/lib",
+	"editor.formatOnSave": true,
+	"[typescript]": {
+		"editor.defaultFormatter": "esbenp.prettier-vscode"
+	},
 	"deno.enablePaths": ["./_deno"],
 	"deno.enable": false
 }
