+++ packages/adapter/adapter-node/src/http/native-fetch.ts
@@ -1,0 +1,3 @@
+import "../polyfills/native-fetch";
+
+export * from "./common";
