+++ packages/adapter/adapter-node/src/http/whatwg-node.ts
@@ -1,0 +1,3 @@
+import "../polyfills/whatwg-node";
+
+export * from "./common";
