+++ packages/adapter/adapter-node/src/http2/index.ts
@@ -1,0 +1,3 @@
+import "../polyfills/node-fetch";
+
+export * from "./common";
