+++ packages/adapter/adapter-node/src/polyfills/node-fetch.ts
@@ -1,0 +1,7 @@
+import installNodeFetch from "@hattip/polyfills/node-fetch";
+import installGetSetCookie from "@hattip/polyfills/get-set-cookie";
+import installCrypto from "@hattip/polyfills/crypto";
+
+installNodeFetch();
+installGetSetCookie();
+installCrypto();
