+++ packages/adapter/adapter-node/src/polyfills/whatwg-node.ts
@@ -1,0 +1,8 @@
+// TODO: Remove or update this rule!
+import installCrypto from "@hattip/polyfills/crypto";
+import installGetSetCookie from "@hattip/polyfills/get-set-cookie";
+import installWhatwgNodeFetch from "@hattip/polyfills/whatwg-node";
+
+installWhatwgNodeFetch();
+installGetSetCookie();
+installCrypto();
