+++ packages/adapter/adapter-node/tsup.config.ts
@@ -3,9 +3,12 @@ import { defineConfig } from "tsup";
 export default defineConfig([
 	{
 		entry: [
-			"./src/index.ts",
-			"./src/native-fetch.ts",
-			"./src/whatwg-node.ts",
+			"./src/http/index.ts",
+			"./src/http/native-fetch.ts",
+			"./src/http/whatwg-node.ts",
+			"./src/http2/index.ts",
+			"./src/http2/native-fetch.ts",
+			"./src/http2/whatwg-node.ts",
 			"./src/request.ts",
 			"./src/response.ts",
 			"./src/fast-fetch.ts",
